move some shared utility methods to EsqlTestUtils
